refactor: improve feature (#145)
* test: ensure errors propagated (#143)

* test: relocate integration tests (#144)

* test: relocate integration tests

* ci: reduce run time

* refactor: remove unused import

* refactor: improve feature isolation

* refactor: remove unnecessary path prefixes

* refactor: invert feature

Using `--features=no_telemetry` still resulted in inclusion of pop-telemetry in the dependency tree. Using `telemetry` also makes conditional compilation much simpler.
