+++ .github/workflows/build.yml
@@ -11,58 +11,68 @@ env:
   GITHUB_ACTOR: pop-cli
 
 jobs:
-  build:
-
+  lint:
     runs-on: ubuntu-latest
-
     steps:
-    - name: Free up space on runner
-      run: |
-        sudo rm -rf /usr/share/dotnet
-        sudo rm -rf /opt/ghc
-        sudo rm -rf "/usr/local/share/boost"
-        sudo rm -rf "$AGENT_TOOLSDIRECTORY"
+      - uses: actions/checkout@v4
 
-    - name: Rust Cache
-      uses: Swatinem/rust-cache@v2.5.0
-      with:
-        cache-on-failure: true
-        cache-all-crates: true
-            
-    - name: Setup git config
-      run: |
-        git config --global user.name "${GITHUB_ACTOR}"
-        git config --global user.email "${GITHUB_ACTOR}@users.noreply.github.com"
+      - name: Check formatting
+        run: cargo fmt --all -- --check
 
-    - name: Install Protoc
-      uses: arduino/setup-protoc@v1.3.0
-      with:
-        version: "3.6.1"
+  build:
+    needs: lint
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
 
-    - name: Rust Setup
-      uses: actions-rs/toolchain@v1
-      with:
-        profile: minimal
-        toolchain: stable
-        target: wasm32-unknown-unknown
-        override: true
-        components: rustfmt, clippy, rust-src
+      - uses: "./.github/actions/init"
+        with:
+          git-user: ${{ env.GITHUB_ACTOR }}
 
-    - uses: actions/checkout@v4
-    - name: Check formatting
-      run: cargo fmt --all -- --check
-        
-    - uses: actions/checkout@v3
-    - name: Check Feature Contracts Excl.
-      run: cargo check --no-default-features --features contract
+      - name: Check Feature Contracts Excl.
+        run: cargo check --no-default-features --features contract
 
-    - uses: actions/checkout@v3
-    - name: Check Features Parachain Excl.
-      run: cargo check --no-default-features --features parachain
+      - name: Check Features Parachain Excl.
+        run: cargo check --no-default-features --features parachain
 
-    - uses: actions/checkout@v3
-    - name: Build default features
-      run: cargo build
+      - name: Build default features
+        run: cargo build
 
-    - name: Run tests
-      run: cargo test --all-features
+  unit-tests:
+    needs: lint
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+
+      - uses: "./.github/actions/init"
+        with:
+          git-user: ${{ env.GITHUB_ACTOR }}
+
+      - name: Run unit tests
+        run: cargo test --lib
+
+  contract-integration-tests:
+    needs: lint
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+
+      - uses: "./.github/actions/init"
+        with:
+          git-user: ${{ env.GITHUB_ACTOR }}
+
+      - name: Run integration tests
+        run: cargo test --test contract
+
+  parachain-integration-tests:
+    needs: lint
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+
+      - uses: "./.github/actions/init"
+        with:
+          git-user: ${{ env.GITHUB_ACTOR }}
+
+      - name: Run integration tests
+        run: cargo test --test parachain
