+++ crates/pop-cli/Cargo.toml
@@ -36,27 +36,22 @@ dirs = { workspace = true, optional = true }
 git2.workspace = true
 
 # telemetry
-pop-telemetry = { path = "../pop-telemetry" }
+pop-telemetry = { path = "../pop-telemetry", optional = true }
 
 [dev-dependencies]
 assert_cmd = "2.0.14"
 predicates = "3.1.0"
 
 [features]
-default = ["contract", "parachain"]
+default = ["contract", "parachain", "telemetry"]
 static-ssl = ["git2/vendored-openssl"]
 contract = [
     "dep:pop-contracts",
     "dep:sp-core",
     "dep:sp-weights",
 ]
-e2e_contract = []
-unit_contract = []
 parachain = [
     "dep:pop-parachains",
     "dep:dirs",
 ]
-e2e_parachain = []
-unit_parachain = []
-no_telemetry = []
-
+telemetry = ["dep:pop-telemetry"]
