+++ crates/pop-contracts/Cargo.toml
@@ -26,7 +26,3 @@ contract-extrinsics.workspace = true
 
 [dev-dependencies]
 tempfile.workspace = true
-
-[features]
-e2e_contract = []
-unit_contract = []
