+++ crates/pop-parachains/Cargo.toml
@@ -29,7 +29,3 @@ walkdir.workspace = true
 # Zombienet
 zombienet-sdk.workspace = true
 zombienet-support.workspace = true
-
-[features]
-e2e_parachain = []
-unit_parachain = []
