+++ crates/pop-parachains/src/lib.rs
@@ -12,7 +12,7 @@ pub use build::build_parachain;
 pub use new_pallet::{create_pallet_template, TemplatePalletConfig};
 pub use new_parachain::instantiate_template_dir;
 pub use templates::{Config, Provider, Template};
-pub use up::{Status, Zombienet};
+pub use up::{Source, Status, Zombienet};
 pub use utils::git::{Git, GitHub, Release};
 pub use utils::pallet_helpers::resolve_pallet_path;
 // External exports
